+++ erc20-metadata/src/calls.rs
@@ -72,22 +72,3 @@ pub fn batch_decimals(contracts: Vec<Address>) -> HashMap<Address, BigInt> {
     }
     results
 }
-
-pub fn batch_total_supply(contracts: Vec<Address>) -> HashMap<Address, BigInt> {
-    let mut results: HashMap<Address, BigInt> = HashMap::new();
-    for chunks in contracts.chunks(CHUNK_SIZE) {
-        let batch = chunks.iter().fold(RpcBatch::new(), |batch, address| {
-            batch.add(nfts::boredapeyachtclub::functions::TotalSupply {}, address.to_vec())
-        });
-        let responses = batch.execute().expect("failed to execute functions::TotalSupply RpcBatch").responses;
-        for (i, address) in chunks.iter().enumerate() {
-            // TotalSupply
-            if let Some(value) = RpcBatch::decode::<BigInt, nfts::boredapeyachtclub::functions::TotalSupply>(&responses[i]) {
-                results.insert(address.to_vec(), value);
-            } else {
-                substreams::log::info!("Failed to decode functions::TotalSupply for address={:?}", Hex::encode(address));
-            }
-        }
-    }
-    results
-}
