+++ erc20-supply/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "erc20-supply"
+description = "ERC20 supply"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+authors = ["Denis <denis@pinax.network>"]
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+substreams = { workspace = true }
+substreams-ethereum = { workspace = true }
+substreams-abis = { workspace = true }
+common = { path = "../common" }
+proto = { path = "../proto" }
