+++ erc20-supply/Makefile
@@ -1,0 +1,19 @@
+ENDPOINT ?= eth.substreams.pinax.network:443
+ROOT_DIR := $(shell dirname $(realpath $(firstword $(MAKEFILE_LIST))))
+SINK_RANGE := ":"
+
+.PHONY: build
+build:
+	cargo build --target wasm32-unknown-unknown --release
+
+.PHONY: pack
+pack: build
+	substreams pack
+
+.PHONY: noop
+noop: build
+	substreams-sink-noop $(ENDPOINT) substreams.yaml map_events -H "X-Sf-Substreams-Parallel-Jobs: 2000"
+
+.PHONY: gui
+gui: build
+	substreams gui -e $(ENDPOINT) substreams.yaml map_events -s 22423815
