+++ erc20-supply/README.md
@@ -1,0 +1,13 @@
+# ERC721 metadata
+
+> `IERC721Metadata.sol`: Interface for the ERC721 metadata extension.
+
+## Includes
+
+- name
+- symbol
+- tokenURI
+
+## References
+
+- https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/token/ERC721/extensions/IERC721Metadata.sol
