+++ proto/src/pb/evm.erc20.metadata.v1.rs
@@ -6,9 +6,6 @@ pub struct Events {
     /// <https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/token/ERC20/extensions/IERC20Metadata.sol>
     #[prost(message, repeated, tag="1")]
     pub metadata_by_contracts: ::prost::alloc::vec::Vec<MetadataByContract>,
-    /// <https://github.com/OpenZeppelin/openzeppelin-contracts/blob/e34251682bac9c3252af30e91e999f13dd098b9f/contracts/token/ERC20/IERC20.sol#L27>
-    #[prost(message, repeated, tag="2")]
-    pub total_supply_by_contracts: ::prost::alloc::vec::Vec<TotalSupplyByContract>,
 }
 #[allow(clippy::derive_partial_eq_without_eq)]
 #[derive(Clone, PartialEq, ::prost::Message)]
@@ -25,15 +22,5 @@ pub struct MetadataByContract {
     /// name() -> string
     #[prost(string, optional, tag="4")]
     pub name: ::core::option::Option<::prost::alloc::string::String>,
-}
-#[allow(clippy::derive_partial_eq_without_eq)]
-#[derive(Clone, PartialEq, ::prost::Message)]
-pub struct TotalSupplyByContract {
-    /// log.address
-    #[prost(bytes="vec", tag="1")]
-    pub contract: ::prost::alloc::vec::Vec<u8>,
-    /// totalSupply() -> uint256
-    #[prost(string, tag="2")]
-    pub total_supply: ::prost::alloc::string::String,
 }
 // @@protoc_insertion_point(module)
