+++ proto/src/pb/evm.erc20.supply.v1.rs
@@ -1,0 +1,20 @@
+// @generated
+// This file is @generated by prost-build.
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct Events {
+    /// <https://github.com/OpenZeppelin/openzeppelin-contracts/blob/e34251682bac9c3252af30e91e999f13dd098b9f/contracts/token/ERC20/IERC20.sol#L27>
+    #[prost(message, repeated, tag="2")]
+    pub total_supply_by_contracts: ::prost::alloc::vec::Vec<TotalSupplyByContract>,
+}
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct TotalSupplyByContract {
+    /// log.address
+    #[prost(bytes="vec", tag="1")]
+    pub contract: ::prost::alloc::vec::Vec<u8>,
+    /// totalSupply() -> uint256
+    #[prost(string, tag="2")]
+    pub total_supply: ::prost::alloc::string::String,
+}
+// @@protoc_insertion_point(module)
